refactor: divided to timeline/Kind
